Beginning support for local bindings, `let _ = _ in _`
Required a bit of restructuring, enabled simple expression parsing as
a result and a primitive notion that a collection of top-level
bindings constitutes a "module".  This is a bit of a quick fix but it
will work for now since I don't (yet) want to work seriously on module
ideas here, first class or otherwise.

No formatting support yet, that's next.